+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## [0.1.1] - 2022-11-09
 
 -   Kubernetes tolerations used in `pyspark-iris-running-on-gke` to make it even more cost-effeective
+-   `pyspark-iris-running-on-gcp-dataproc-serverless` starter
 
 ## [0.1.0] - 2022-10-24